pyproject.toml
[tool.poetry]
name = "cricinfo_app"
version = "0.1.0"
description = "cricket scrore by day"
authors = ["Neeraj Tickoo <ntickoo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
django-livereload-server = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
